rootfs: change order of onboot containers and do modprobe after storage-init
storage-init does not (and should not) depend on any extra loaded modules.
Loading of all extra modules (network, leds, usb, etc) is the last operation
of the whole boot process.

This behavior becomes also important if we have a dump-capture kernel (future
patches), which prevents modules load and relies only on what is builtin.
The behaviour of no-module-load is crucial for dump-capture kernel stability.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>