pillar: introduce a new field DatastoreIDsList
Introduce a new field 'DatastoreIDsList' which holds list of all
datastores IDs, which can be used for downloading an image in
case of network failure.

Many diff lines in the structure definetions are caused by structure
member alignments.

In this patch a new field is populated and copied between structures,
but does not replace the original DatastoreID field. This will happen
in the following patches.

The following structures are affected:

   ContentTreeConfig
   ContentTreeStatus
   DownloaderConfig
   DownloaderStatus
   BlobStatus

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>