volumemgr: start using the 'FallbackDatastoreIDsList' field
ContentTreeStatus now has the list of datastore types in order
to keep track of datastore updates, which appear out of sync with
the content tree. There are two async code paths:

  o Once a new ContentTreeStatus is created, we do a lookup over
    the datastores, which IDs are taken from the DatastoreID and
    the list of fallback IDs. If all datastores are found, the
    whole ContentTreeStatus is marked as 'AllDatastoresResolved'.
    If not all the datastores are resolved - an update from a
    datastore is awaited.

  o On each update from a datastore we do a lookup by ID and
    account resolved datastores. If all datastores are resolved
    the whole structure is marked as resolved.

The 'IsOCIRegistry' field is special. The logic of handing the OCI
registry is quite complicated and can't be easily adopted for the
list of the datastores, so for now I expect that OCI registry is
used only for the singleton datastore, namely when fallback list
is empty.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>