pillar: replace all uses of the 'DatastoreID' with the 'DatastoreIDsList[0]'
Unfortunately this is a hacky approach and is used only
for the purpose not to break the existing code: we still
depend on the singleton datastore ID field, albeit now
we have a list of datastores.

The main problem is that the singleton datastore ID is
used along the whole downloader stack in different cases
and for now I do not see a simple way how to use the list.

I keep this patch separate to highlight these places in the code,
that need to be refactored.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>