zedagent: define a destination type of the "info" message
For now we have two destinations for the "info" message: actual
controller (cloud) and local server. By default "info" message
is not periodic, but is sent only on update, which causes a
problem for the stateless LOC implementation (not part of this
patch set).

The idea of the 'infoDest' enum is to have a controller and a
local server be served as usual - by the "info" messages
updates, and only LOC should receive periodic "info" messages.
In order to send periodic "info" messages only to LOC we need
to define destination type (the purpose of this patch).

This patch defines a type for different destinations and a
function 'infoDestToURLs()' which provides a convenient way
to get a list of destination URLs according to the destination
type passed as a parameter.

In the following patches type will be used for all Publish*()
calls.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>