pkg/pillar/go.mod,go.sum: update lf-edge/eve/libs module to use the latest
This repeats the same trick used for the lf-edge/eve/api/go: instead
of putting the latest SHA we replace it with a local implementation.
Unfortunately golang still uses the vendor copy, but at least we
don't need to update the go.mod and go.sum manually in order to
reflect latest changes in the eve/libs, so what is needed after
the change in eve/libs/ is the following:

  1. make proto-vendor
  2. git add -u
  3. git commit
  4. PROFIT

Why all this mess? There was a commit to the libs/zedUpload folder:

  23cea8b9b948 ("zedUpload/http: fail faster for not successfull HTTP requests")

which did changes in the libs, but changes to the vendor were forgotten
and not committed. So to minimize such a nasty mistake it would be
nice to make this dependency somehow automated. This commit does not
automate, but at least gets rid of the extra body movement for copying
latest SHA to the go.mod file, making it less error prone.

The bright future probably comes with the next commit. Stay tuned.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>