http: revert the 5352c7d6c461 ("http: minimize timeout by propagating error")
The patch reverts the commit, which was an attempt to make fail
faster. The problem with the commit is that it does not retry
the download if it was timeouted. The outher problem is the
resource leak: the resp.Body should be closed. The resource
leak will be targeted in the next commit.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>