pillar/scripts/device-steps.sh: switch to chronyd from busybox ntpd
Use chronyd for time synchchronization instead of busybox ntpd
because of the rich set of features it has and monitor API
implemented in golang, which we can use for getting all the
useful information for ntp status sharing to the controller.

The only difference between how ntpd and chronyd starts is the
way how an NTP server/peer should be passed to the daemon.
In case of chronyd the chrony client (chronyc) should be used.

Also now the default pool.ntp.org is explicitly passed to the
chronyd as 'pool', other NTP servers which come from the nim
passed as 'server'.

In this change all common code was moved to the functions,
namely single_ntp_sync(), start_ntp_daemon() and
restart_ntp_daemon(). This reduces code duplication.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>