zedcloud/deferred: respect URL argument as a key on starting new deferred
Since a while we support a few destinations: Controller, LOC, LPS.
There are a few calls which send messages to all destinations
simultaneiously. The semantics of the `SetDeferred` function is to
replace a deferred item if key is found. The `queueInfoToDest()`
can add two items to the same `DeferredPeriodicCtx` context with
different URLs. The original behavior of the `SetDeferred` call
is to replace first item with the second item by the key, which
will be same. This leads to a missing packet to one of the URLs
if Controller and LOC are both enabled.

This patch adds `URL` check along with the `key` check on deferred
item add, makes both items with different URLs co-exist in the
same queue.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>